+++ src/components/common/container/index.module.css
@@ -1,6 +1,6 @@
 .container {
 	width: 475px;
-	padding: 5px;
+	padding: 10px;
 	border-top: 1px solid #e7e7e7;
 	border-right: 1px solid #e7e7e7;
 	border-left: 1px solid #e7e7e7;
