+++ src/components/cookingInstructionCard/index.module.css
@@ -6,5 +6,5 @@
 .title {
 	font-size: 25px;
 	margin: 0;
-	margin-top: 20px;
+	margin-top: 30px;
 }
