+++ src/components/cookingInstructionItem/index.module.css
@@ -2,7 +2,7 @@
 	background-color: rgb(246, 241, 152);
 	border-radius: 12px;
 	padding: 10px 15px;
-	margin: 5px 0 10px 0;
+	margin: 10px 0 15px 0;
 }
 
 .title {
