+++ src/components/popularAndReviewCard/index.module.css
@@ -1,46 +1,31 @@
-.container {
-	position: relative;
+.subtitle {
+	font-size: 20px;
+}
+
+.recipeWrapper {
 	display: flex;
-	margin: 0 calc(100% - 83%);
+	padding: 0;
+	overflow-x: scroll;
 }
 
-.imageWrapper {
-	width: 160px;
-	height: auto;
-}
-
-.img {
-	width: 100%;
-	height: 140px;
-	border-radius: 20px;
-}
-
-.infoWrapper {
-	display: flex;
-	flex-direction: column;
-	justify-content: space-evenly;
-	width: 180px;
-	font-size: 12px;
-	padding: 0 12px;
-	margin-bottom: 5px;
-}
-
-.title {
-	font-size: 15px;
-	font-weight: 600;
-	margin-bottom: 3px;
-}
-
-.descWrapper {
-	display: flex;
-}
-
-.iconWrapper {
+.subtitleWrapper {
 	display: flex;
 	align-items: center;
-	margin-right: 3px;
+	justify-content: space-between;
+	margin-right: 10px;
 }
 
-.desc {
-	margin-right: 12px;
+.see_all {
+	color: #ffffff;
+	font-size: 14px;
+	border: none;
+	border-radius: 20px;
+	padding: 6px;
+	opacity: 90%;
+	background-color: #f7a78ed1;
+	cursor: pointer;
+}
+
+.see_all:hover {
+	opacity: 80%;
 }
