+++ src/components/todayRecipeCard/index.module.css
@@ -1,0 +1,15 @@
+.container {
+	margin: 30px 0 20px 0;
+}
+
+.title {
+	margin: 0;
+	font-size: 20px;
+}
+
+.recipeWrapper {
+	display: flex;
+	padding: 0;
+	overflow-x: scroll;
+	/* overflow: hidden; */
+}
