+++ src/pages/index.module.css
@@ -1,41 +1,9 @@
 .title {
-	font-size: 20px;
-	margin: 5px 0 10px 0;
-}
-
-.subtitle {
-	font-size: 18px;
+	font-size: 22px;
+	margin: 15px 0 10px 0;
 }
 
 .strong {
 	color: #ffb823;
 	font-size: 24px;
-}
-
-.recipeWrapper {
-	display: flex;
-	padding: 0;
-	overflow-x: scroll;
-	/* overflow: hidden; */
-}
-
-.subtitleWrapper {
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	margin-right: 10px;
-}
-
-.see_all {
-	color: #ffffff;
-	border: none;
-	border-radius: 20px;
-	padding: 4px;
-	opacity: 90%;
-	background-color: #f7a68e;
-	cursor: pointer;
-}
-
-.see_all:hover {
-	opacity: 80%;
 }
