+++ src/pages/recipe/all/index.module.css
@@ -1,6 +1,7 @@
 .container {
 	display: flex;
 	justify-content: center;
-	height: calc(100vh - 150px);
+	margin-top: 10px;
+	height: calc(100vh - 115px);
 	overflow-y: scroll;
 }
